+++ maiacore/src/maiacore/python_wrapper/py_clef.cpp
@@ -7,11 +7,17 @@
 namespace py = pybind11;
 
 void ClefClass(const py::module& m) {
+    py::enum_<ClefSign>(m, "ClefSign")
+        .value("G", ClefSign::G)
+        .value("F", ClefSign::F)
+        .value("C", ClefSign::C)
+        .value("P", ClefSign::PERCUSSION);
+
     m.doc() = "Clef class binding";
 
     // bindings to Clef class
     py::class_<Clef> cls(m, "Clef");
-    cls.def(py::init<const std::string&, int>(), py::arg("sign") = "G", py::arg("line") = 2);
+    cls.def(py::init<const ClefSign, int>(), py::arg("sign") = ClefSign::G, py::arg("line") = -1);
 
     cls.def("getLine", &Clef::getLine);
     cls.def("setLine", &Clef::setLine, py::arg("line"));
@@ -19,11 +25,13 @@ void ClefClass(const py::module& m) {
     cls.def("getSign", &Clef::getSign);
     cls.def("setSign", &Clef::setSign, py::arg("sign"));
 
+    cls.def("toXML", &Clef::toXML, py::arg("clefNumber") = -1, py::arg("identSize") = 2);
+
     // Default Python 'print' function:
-    cls.def("__repr__", [](const Clef& clef) { return "<Clef " + clef.getSign() + ">"; });
+    cls.def("__repr__", [](const Clef& clef) { return "<Clef " + clef.getClefSignStr() + ">"; });
 
     cls.def("__hash__", [](const Clef& clef) {
-        return std::hash<std::string>{}(clef.getSign() + std::to_string(clef.getLine()));
+        return std::hash<std::string>{}(clef.getClefSignStr() + std::to_string(clef.getLine()));
     });
 
     cls.def("__sizeof__", [](const Clef& clef) { return sizeof(clef); });